refactor(tabbar): split behavior into hooks and components
