+++ src/components/layout/tabbar/constants.ts
@@ -1,0 +1,46 @@
+export const RESOURCE_I18N_KEYS: Record<string, [string, string]> = {
+  "cluster-overview": ["cluster", "overview"],
+  "resource-diagram": ["cluster", "resourceDiagram"],
+  nodes: ["cluster", "nodes"],
+  events: ["cluster", "events"],
+  namespaces: ["cluster", "namespaces"],
+  leases: ["cluster", "leases"],
+  "helm-releases": ["helm", "releases"],
+  "flux-kustomizations": ["flux", "kustomizations"],
+  "workloads-overview": ["workloads", "overview"],
+  deployments: ["workloads", "deployments"],
+  pods: ["workloads", "pods"],
+  replicasets: ["workloads", "replicaSets"],
+  daemonsets: ["workloads", "daemonSets"],
+  statefulsets: ["workloads", "statefulSets"],
+  jobs: ["workloads", "jobs"],
+  cronjobs: ["workloads", "cronJobs"],
+  "port-forwards": ["networking", "portForwards"],
+  services: ["networking", "services"],
+  ingresses: ["networking", "ingresses"],
+  "endpoint-slices": ["networking", "endpointSlices"],
+  "network-policies": ["networking", "networkPolicies"],
+  "ingress-classes": ["networking", "ingressClasses"],
+  secrets: ["configuration", "secrets"],
+  configmaps: ["configuration", "configMaps"],
+  hpa: ["configuration", "hpa"],
+  "limit-ranges": ["configuration", "limitRanges"],
+  "resource-quotas": ["configuration", "resourceQuotas"],
+  "pod-disruption-budgets": ["configuration", "podDisruptionBudgets"],
+  "persistent-volumes": ["storage", "persistentVolumes"],
+  "persistent-volume-claims": ["storage", "persistentVolumeClaims"],
+  "volume-attachments": ["storage", "volumeAttachments"],
+  "storage-classes": ["storage", "storageClasses"],
+  "csi-drivers": ["storage", "csiDrivers"],
+  "csi-nodes": ["storage", "csiNodes"],
+  "service-accounts": ["accessControl", "serviceAccounts"],
+  roles: ["accessControl", "roles"],
+  "role-bindings": ["accessControl", "roleBindings"],
+  "cluster-roles": ["accessControl", "clusterRoles"],
+  "cluster-role-bindings": ["accessControl", "clusterRoleBindings"],
+  crds: ["administration", "crds"],
+  "priority-classes": ["administration", "priorityClasses"],
+  "runtime-classes": ["administration", "runtimeClasses"],
+  "mutating-webhooks": ["administration", "mutatingWebhooks"],
+  "validating-webhooks": ["administration", "validatingWebhooks"],
+};
