+++ src/components/layout/tabbar/index.ts
@@ -1,0 +1,1 @@
+export { TabBar, useTabTitle } from "./TabBar";
